nits here and there
