+++ src/utils/transactionUtils.ts
@@ -1,0 +1,25 @@
+import type { Event } from "@/client";
+
+export function formatDate(dateString?: string) {
+  if (!dateString) return "";
+  const date = new Date(dateString);
+  return date
+    .toLocaleDateString("en-US", {
+      year: "numeric",
+      month: "short",
+      day: "2-digit",
+    })
+    .toUpperCase();
+}
+
+export function groupByDate(transactions: Event[]) {
+  return transactions.reduce(
+    (acc, tx) => {
+      const date = formatDate(tx.createdAt);
+      if (!acc[date]) acc[date] = [];
+      acc[date].push(tx);
+      return acc;
+    },
+    {} as Record<string, Event[]>,
+  );
+}
